Moved data transformations to own file
