+++ src/app/transactions/page.js
@@ -31,6 +31,7 @@ import {
     TransactionsTableDeleteModalDismissButton
 } from './actions/delete-modal';
 import CsvPreviewTable from './csv-preview-table';
+import { transformCsvData } from '@/utility/transformers';
 import Papa from 'papaparse';
 
 const columnHelper = createColumnHelper();
@@ -111,29 +112,6 @@ const Transactions = () => {
 
         toggleTables()
         setCsvData([]);
-    }
-
-    function transformCsvData(csvData) {
-        const newData = [];
-        for (const row of csvData) {
-            newData.push({
-                amount: row['Amount'],
-                balance: row['Balance'],
-                checkNumber: row['Check Number'],
-                description: row['Description'],
-                effectiveDate: row['Effective Date'],
-                extendedDescription: row['Extended Description'],
-                memo: row['Memo'],
-                postingDate: row['Posting Date'],
-                referenceNumber: row['Reference Number'],
-                transactionCategory: row['Transaction Category'],
-                transactionId: row['Transaction ID'],
-                transactionType: row['Transaction Type'],
-                type: row['Type']
-            })
-        }
-
-        return newData;
     }
 
     async function handleUpload() {
